+++ drizzle/0082_pinned_notes.sql
@@ -1,0 +1,1 @@
+ALTER TABLE `diary_entries` ADD `pinned_at` text;